s2s: add frozen coordinates func towards TS drivers for filtered mode
Allow a call to check whether coordinates should be retargetted in
Touchscreen drivers to a frozen X/Y and let TS driver pass real coords
for s2s checking. This is to avoid framework user land miss events
and occasionally crash if not prepared for missing X/Y pos events.